+++ CHANGES.md
@@ -1,11 +1,19 @@
 # CHANGES
 
+#v0.1
+## v0.1.0
+* created client-side FreeboxFinder, Location
+* moved most of stuff from `scripts.js` to `freebox-finder.js`
+* created config directory for config files
+* now using jade for basic layout
+* created a views directory
+
 # v0.0
 ## v0.0.1
-* created package.json
-* created TODO.md
-* created CHANGES.md
-* created config.default.json
+* created `package.json`
+* created `TODO.md`
+* created `CHANGES.md`
+* created `config.default.json`
 * defined FreeboxFinder class - extends eventEmitter
 * defined boxSchema
 * setup mongoose connection